Bug 1194905 - Build libvpx neon code without -mthumb and -mfloat-abi=softfp
-mfloat-abi=softfp is hardcoding an ABI that may not match everything else
in Gecko, which is bound to fail. It's also unrelated to building NEON code.
-mthumb is also unrelated to building NEON code, and should be left as it
is for everything else.
Now, interestingly, we also have a VPX_ASFLAGS variable that just contains
the right things instead of hardcoding them in the moz.build, so just use
that instead.

Gbp-Pq: Topic fixes
Gbp-Pq: Name Bug-1194905-Build-libvpx-neon-code-without-mthumb-an.patch